components/backend/pyproject.toml
[tool.poetry]
name = "procurement-assistant-be"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Ahmed Moorsy <ahmed.moorsy798@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
numpy = "^1.22.4"
motor = "^3.6.0"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"

faiss-cpu = "^1.9.0.post1"
langchain = "^0.3.14"
langchain_core = "^0.3.29"
langchain_openai = "^0.2.14"
langchain-community = "^0.3.14"
langchain-huggingface = "^0.1.2"
langgraph = "^0.2.61"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
